+++ examples/record-playback/CMakeLists.txt
@@ -7,7 +7,7 @@ project(RealsenseExamplesRecord-Playback)
 
 
 if(BUILD_GRAPHICAL_EXAMPLES)
-    add_executable(rs-record-playback rs-record-playback.cpp ../example.hpp ../../third-party/imgui/imgui.cpp ../../third-party/imgui/imgui_draw.cpp ../../third-party/imgui/imgui_impl_glfw.cpp ../../common/fs.cpp)
+    add_executable(rs-record-playback rs-record-playback.cpp ../example.hpp ../../third-party/imgui/imgui.cpp ../../third-party/imgui/imgui_draw.cpp ../../third-party/imgui/imgui_impl_glfw.cpp ../../common/filesystem.cpp)
     set_property(TARGET rs-record-playback PROPERTY CXX_STANDARD 11)
     target_link_libraries(rs-record-playback ${DEPENDENCIES})
     include_directories(../ ../../third-party/tclap/include ../../third-party/imgui)
